[BUGFIX] Remove EXT:backend dependencies from errorpage.css
In #95372 the errorpage.css, used for the standalone
ErrorPage template, was moved from EXT:core to
EXT:backend, since it had dependencies to this
extension. However, it actually belongs to EXT:core.
The ErrorPageController is used for both, backend
and frontend.

A side effect of the relocation to EXT:backend: The
errorpage.css was now always loaded in the backend,
overriding a couple of classes, which then led to
display errors.

This is now fixed by removing any dependency to
EXT:backend and moving the stylesheet back to
EXT:core.

As a side effect, the "typo3_orange.svg" file
is removed from EXT:backend and EXT:install,
since it's now available in EXT:core. All
places are adjusted accordingly.

Resolves: #95388
Related: #95372
Releases: master
Change-Id: I697aa0f98dbd989f68779690d466c7bfdce60c2b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71343
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>