[BUGFIX] Incomplete setup in TcaCategoryTest
Some tests trigger access to $result['rootline'] and
$result['site']. In normal usages, those are set up by
the FormEngine $result defaults, which is not applied
if calling a data provider directly as done with the
functional test in question. Solution in to set the
values to defaults within tests.
This fixes a currently suppressed PHP 8.0 warning for
functional tests.

Resolves: #95504
Releases: master
Change-Id: Ia0c8ce05a94e3ef3c731c933469fb648c779017a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71472
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>