[BUGFIX] Restore inconsistent language highlighting
The page module displays a flash message for page
translations with inconsistent ("mixed mode") content
elements. The last sentence of the flash message reads:
"We highlighted the problematic records for you". This
however was no longer the case in the fluid based page
module.

Therefore, this patch restores highlighting of such records.

Resolves: #95717
Related: #91177
Releases: master, 10.4
Change-Id: Ic0ada77428a426254f9cbba256e23f938733f9aa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72031
Tested-by: Tobias Gaertner <tobias.gaertner@benaja-websolutions.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Tobias Gaertner <tobias.gaertner@benaja-websolutions.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>