[TASK] Ensure record_commit parameters are not reused
When using the "saveAndClose" option for a newContentElementWizard
item, the target link forwards directly to the record_commit route
instead of FormEngine. Previously the parameters for this link
were added to an array, which was then never reset after used
for the link creation. This therefore resulted in multiple
elements were created as soon as the "saveAndClose" option
was set multiple times.

This is now fixed by not using a single use variable anymore,
but directly building the data array in the link, like it's
already done for the FormEngine links.

Releases: master
Resolves: #95526
Related: #95277
Change-Id: I897acd376ce44c6f26c27dc7ddd927302e5be8bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71488
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>