[BUGFIX] Enforce setCopyMode in clipboard
The clipboard component features helper methods,
which generate a URL to update the clipboard
state with new elements. Those are used in
recordlist and filelist.

Previously the "setCopyMode" parameter was
only set in case the element should be "copied".
This then might lead to no "mode" set in the
clipboardData at all. Like other modules
already do, the helper methods do now always
set the "setCopyMode" parameter to enforce
the generation of the "mode" key in the
clipboardData.

Anyways, an access of the "mode" in the page
module is secured by a null collapse operator.

Resolves: #95536
Releases: master
Change-Id: I0799f711cd3ab8b6be702b62ba26d974d46da2d0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71504
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>