[BUGFIX] Use CodeMirror auto sizing feature correctly
To use CodeMirror's auto sizing feature, we need
to set the style attribute of the wrapper element
to "auto", instead of providing the "height" option.

In a previous patch a related issue was fixed by
setting the height to the remaining space of the
frame. This led to another issue, which occurs,
when there is already little space left.

Both is fixed by using the auto sizing option
correctly and even gives us the benefit of
automatic resizing, when changing the editor's
content.

Resolves: #95801
Related: #86699
Releases: master
Change-Id: I856aa85dfac348db1a49d9c73306159e145f9eeb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71998
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>