[BUGFIX] Fix array access for user without read permissions in PHP8
Given you have a user which has no access to the current page
or no page at all, a "access array offset on value of type"
is thrown in PHP8.

This has been fixed by adding a fallback in case the
array key does not exist which then throws the expected
exception.

Resolves: #96026
Releases: master, 11.5
Change-Id: I335a74fb6c2566d7a3e65632977dbc2740249414
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72228
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>