[BUGFIX] Make belayoutwizard TSconfig preview readonly
With #94222, the TSconfig preview of a TCA field with
renderType= belayoutwizard was adjusted to use the
CodeMirror plugin - in case EXT:t3editor is installed.

With this change, also the default "preview" was changed
form a <code> element to a <textarea>. This therefore led
to the field being editable. However, any manual change
will always be overwritten on adjusting the grid via the editor
or on saving the record.

To avoid confusion for the user and to be in line with
the CodeMirror variant, the previous behaviour is restored
by adding the readonly attribute to the <textarea> element.

Resolves: #96064
Related: #94222
Releases: master, 11.5
Change-Id: Iaa41df59cbfd0420962c2239a42d329c249adb6c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72267
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>