[BUGFIX] Make selectCheckBox work with readOnly
Using "readOnly" for a TCA field with renderType
"selectCheckBox" previously led to the fact, that
nothing was rendered at all. This can be seen with
`tx_styleguide_elements_select.select_checkbox_4`.

This is now fixed, while the whole element got a
light code cleanup. This will increase performance
a bit, since unneeded JavaScript is no longer loaded.

Additionally, this also fixes the "expandAll" option,
which was broken due to the bootstrap 5 upgrade.

Resolves: #96058
Resolves: #96068
Releases: master, 11.5
Change-Id: I762d328c9f4c83558dd8fd98683f8ccfe6c4b3f0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72279
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>