[BUGFIX] Pass default/override values to FormSelectTreeAjaxController
The SelectTree and the CategoryTree use a placeholder
form element with all the information like current value
and then calls the FormSelectTreeAjaxController to
actually build the tree.

When using default / override values for such field,
the information was previously only available on
building the placeholder. Therefore the value was
correctly filled and stored, but not displayed in
corresponding tree.

This is now fixed by passing those values to the
Ajax endpoint, which then adds them to the form
compiler input. This allows the data providers to
properly select the values in the tree.

Resolves: #91607
Releases: main, 11.5
Change-Id: I8743bc3c28977b8d8758096b573a0be332d825e8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72645
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>