[BUGFIX] Fix scoping issue in FormEngineValidation
theChar is not initialized and therefore pollutes global scope.
(It is assigned to window.theChar in the FormEngine iframe).

Currently not an error (we are not using strict mode in RequireJS),
just bad style, but once we switch to strict mode (type="module"),
this will fail without proper declaration.

Releases: main, 11.5
Resolves: #96395
Related: #96323
Change-Id: Ie82da6a1b664e4010179cccaabb07e7d87b19038
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72717
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>