[BUGFIX] Allow CSV in startingPoints config as advertised
The category starting points introduced with #95037 were designed to
allow comma-separated values in its configuration. Due to improper
tests, this specific notation was not discovered as being broken right
now.

If the incoming configuration is a string, GeneralUtility::intExplode()
is applied to remove any non-integer value. Afterwards, a new CSV
string is generated.

Resolves: #96397
Related: #95037
Releases: main, 11.5
Change-Id: I0759d8093bb4665b7709f4e20539307467f246df
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72775
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>