[BUGFIX] Fix ordering of db mounts in page tree
With the changes in #95854, database mounts
are no longer sorted by the order they are defined in
the group/user settings, but undefined by the
order the database returns.

DB mounts aren't sorted against each other,
that means sorting values are not meaningful anyway,
and results in uncontrolable orderings.

Explicitly preserve the order using usort().
Also revert groupAndSortPages() to the state prior
to #95854, as the forcePid parameter would now be unused.

Releases: main, 11.5, 10.4
Resolves: #96359
Related: #95854
Change-Id: I64b8d90c68246cc0ac01abcbf3678aa1055189e3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72854
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>