[BUGFIX] Fix NavigationContainer iframe loading
Some browsers (reported were Safari and Firefox) are not able to
load the file tree when invoking the module the first time.
This is because the iframe src is initially set to about:blank
and ModuleMenu calls both, iframe.setAttribute('src', …) and
iframe.contentWindow.location.reload(). This seems to cause
browsers to abort load as the current URL is being reloaded.

This is actually a result of a refactoring in #52877 where
the condition navUrl != currentUrl previously correctly
issued a `setUrl()` call but afterwards was swapped to
call `refresh()` instead (or rather additionally).

Releases: 10.4
Resolves: #96475
Resolves: #92556
Related: #52877
Change-Id: Ib20c05e0ca6ed1de62cabf86b6a46a3399636fab
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72932
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>