[!!!][TASK] Remove TYPO3_CONF_VARS['SYS']['USdateFormat']
$GLOBALS['TYPO3_CONF_VARS']['SYS']['USdateFormat'] = true/false
has been added in TYPO3 4.0 with a happy "Go US!" message.

It basically sometimes sets date displays to 'MM-DD-Y'
over 'DD-MM-Y' in a couple of views. This is
inflexible and has been superseded at some point by
$GLOBALS['TYPO3_CONF_VARS']['SYS']['ddmmyy'].

The TYPO3 backend is still incapable of rendering dates in
user selected dates and time zones. This patch is a pre-patch
to improve that situation by for now removing and ignoring
the 'USdateFormat' option. This unblocks a rollout of
the more general option 'SYS/ddmmyy' and/or paves the
way to a broader date and timezone related refactoring
that incorparates single user preferences.

Change-Id: I84e43c293c0c2aec86c65ac3716f8b98bee31ac4
Resolves: #96550
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73031
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>