[BUGFIX] Prevent null pointer in PageLayoutView
Under some circumstances, e.g. in a translations
only page module, a given $languageId (in this case
the default `0`) is not part of the $siteLanguages
array, since this only contains languages, allowed
for the user.

To fix a null pointer exception, we check whether
the SiteLanguage exists, before accessing it.

Resolves: #96600
Releases: 10.4
Change-Id: I2897edbf932075645a0c00dddb441b7eaa098708
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73102
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>