[BUGFIX] Hide view page button in special all languages view
The "View webpage" button in the docheader of the
page module allows to view the page in the currently
selected language. This means, the button always
depends on the selected language.

With #92460 the special "all languages" option was
added to the language selector (in languages mode).

Since this option is no valid language id, the button
is now hidden, in case "All languages" is selected.

Resolves: #96603
Releases: main, 11.5
Change-Id: If92b46541868b9cfd1c6aa7addd191d8f877056d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73082
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>