[TASK] Replace method clone with 'GeneralUtility::removeDotsFromTS()'
'EditDocumentController::parseAdditionalGetParameters()'
was implemented to remove dots from key identifiers in
TypoScript array. Using 'callByRef' as parameter argument
should be avoided and is harder to mock and test properly.

This patch replaces this method with the well covered public
'GeneralUtility::removeDotsFromTS()', which is a safe drop-in
replacement. This eliminates the only usage, thus method is
removed along with the corresponding test.

Resolves: #96712
Releases: main
Change-Id: I593414e1ab4735c4b3ff385a8eb8ef99be7a7512
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73252
Tested-by: core-ci <typo3@b13.com>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>