[BUGFIX] Avoid PHP 8.0/8.1 error in RecordHistory
When passing null as $value parameter, an exception is
thrown as the return value must be of type string. We
avoid this by adding string type-hints to the signature
of two methods and string cast the values directly in
the class constructor.

It should be investigated constructor arguments can be
tightened with another patch.

Resolves: #96741
Releases: main, 11.5
Change-Id: I1d589faceb5969c0dee189fbc385cb63a568f3bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73270
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>