[BUGFIX] Add hidden fields when opening collapsed IRRE elements
A record is rendered that has an inline record. That
record is rendered collapsed (only header rendered).
If such a child now has a renderType="hidden" field,
and the record is opened, the hidden field HTML is
missing in the ajax response.
The patch now adds the hidden field html when opening
the record via ajax (trigger renderFieldsOnly).

Resolves: #96781
Related: #87614
Releases: main, 11.5
Change-Id: Ic887712287bd3ce2903e79cc6f44a01f87ad11f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73394
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>