[BUGFIX] Ignore deleted l10n_source records when translating
If l10n_source points to a deleted or non-existing record the
translation wizard failed in free mode with a 500 error,
because determining the source languages fail and no fallback was
implemented. Instead of failing, the behaviour was adjusted to
handle l10n_source pointing to erroneous records similarly to an
empty column and offering all available language as option for
the translation - as in fact the source cannot be reliably determined.

Resolves: #96838
Releases: main, 11.5
Change-Id: Iec7f4a1dba6b93c5185fec7863b1deaa175b0279
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73433
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Benni Mack <benni@typo3.org>