[BUGFIX] Catch non-okayish responses in Modal requests
If the modal component starts an AJAX request that results in a
non-okayish response (>= 400), the underlying fetch() throws an error
that is now caught.

If the response sends an optional body, it is rendered in the modal. If
an empty body is received, a fallback text explaining the issue is
rendered instead.

Resolves: #96852
Releases: main, 11.5
Change-Id: I986c0b92d742fa76b773dd3f3a976a05a3f6b546
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73448
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>