[BUGFIX] Prevent adding invalid LinkButton in edit document
Prior to #96777, the preview url for tor "view" button was
created using BackendUtility::getPreviewUrl), which threw
an exception, in case no link could be created.

With the change to use the PreviewUriBuilder directly,
no such exception is thrown, but buildUri() just returns
NULL. This then lead an invalid link exception, in the
EditDocumentController, since a LinkButton has to
provide a target.

This is now fixed by removing the try/catch and just
checking the generated preview uri for not being empty.

The same is done at the other places, which got adjusted
in the related patch.

Resolves: #96875
Related: #96777
Releases: main
Change-Id: I9041a76a9c89eaf932cf002b26c224f0d97950f4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73470
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>