[BUGFIX] Properly resolve field values in content preview
The StandardContentPreviewRenderer runs
through a list of fields, whose value should
be processed and displayed in the footer.

Some of those field values might be a reference
to a foreign table, e.g. in the "fe_group" field.
Due to changes in getProcessedValue(), it's
necessary to provide the records' uid, to be
able to resolve values of fields with a
"foreign_table" configuration.

Resolves: #96920
Releases: main, 11.5
Change-Id: I0d8739710e478a7cb264be1d5b90b70af31de964
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73563
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>