[BUGFIX] Fix PHP 8.1 warning in StoragePermissionsAspect
BackendUserAuthentication isn't an injectable service,
therefore the (unused) dependency injection is dropped
and $GLOBALS['BE_USER'] usage is moved to be scoped
by the isBackend() check.

Also fix the same DI-anti-pattern in CategoryPermissionsAspect.
No user facing error is currently known for this case,
as the ModifyTreeDataEvent is only executed in backend
context, so this is just to streamline both aspects.

Releases: main, 11.5
Resolves: #96960
Change-Id: I7b9f49ef82eeffcf369f471b5d031311aa5007b4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73544
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>