[BUGFIX] Show CSH popup window without backend frame
Since #94084 all module URLs are automatically framed by
the TYPO3 main controller whenever they are opened
in a Sec-Fetch-Dest aware request.

Use a non-module route for the popup window in order
to avoid the automatic re-framing.

Also hide the back-button in the popup window, as it would
cause the entire backend to be opened in the popup.
The shortcut button will link to the module-based
detail-action in order to show in regular module context.

Releases: main, 11.5
Resolves: #95361
Resolves: #86099
Related: #94084
Change-Id: I0bae3e502fa58f4b2bee804c5ed12b0fc2030d20
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73608
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>