[!!!][FEATURE] Add PSR-14 Events for modifying inline element controls
This introduces two new PSR-14 events, enabling
extension authors to modify the enabled controls as
well as the corresponding controls' markup for
FormEngine inline elements.

This replaces the previously available hook
$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tceforms_inline.php']['tceformsInlineHook']
which is therefore now removed.

The hook required implementations to implement
the InlineElementHookInterface, which is now
deprecated as it's unused due to the hook removal.

Resolves: #97231
Releases: main
Change-Id: I669d759c554d4a2d3ed24f65013c74d54038a0af
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74026
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>