[!!!][TASK] Remove TCA option "cruser_id"
TYPO3's TCA definition has a setting called `cruser_id`
which automatically is filled with the current TYPO3's
Backend User ID, if created via DataHandler.

However, this information is also available in the
RecordHistory (since its extraction from sys_log in v9)
and can be queried, when needed, directly from there.

This way, all TCA-based tables, having the setting "cruser_id"
do not need to carry this value anymore, as it is duplicated
in TYPO3's data model.

This change removes the evaluation of "cruser_id",
and removes the filling of this property via DataHandler.

The DB field is also not set anymore automatically.

The Info popup (ElementInformationController) now
shows the information from the RecordHistory information.

Resolves: #98024
Releases: main
Change-Id: I1fe54b0ecfc0ee9b2d1867011dd960e1906fd2d4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75123
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>